+++ simple-dns/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 - Add RFC 1183
+- Add RFC 1706
 - Add new error type `NoEnoughData` for stream parsing
 
 ## 0.4.6