+++ simple-dns/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add RFC 1183
 - Add RFC 1706
 - Add RFC 1876
+- Add NOTIFY code (RFC 1995 and 1996)
 - Add new error type `InsufficientData` for stream parsing
 
 ## 0.4.6